Added "slow" test (many iterations)
I registered a "slow" mark in pytest.ini and
skip tests with this tag by default.

The test added runs an inversion for a week of data, with 5000
iterations.

I'm not sure how useful it is now, although if this test runs quickly,
then there is probably something wrong with the model.