Site inidicators preserved if data dropped
The site indicators now depend on the order of the sites
in the input to `fixedBasisMCMC`, so if a site is dropped,
the site indicators of the other sites are unchanged.

This means that between runs over different time periods,
site indicators will be consistent.

For instance, if sites = ["MHD", "TAC", "BSD"] and no data
is found for "TAC", then "MHD" will still have site indicator 0
and "BSD" will still have site indicator 2.