Test for reload merged = True, but no merged found
We recently found a bug where if `reload_merged_data=True`, but
no merged data is found, then `data_processing_surface_notracer`
is not run, so `fp_all` isn't defined.

I added a test to make sure this works correctly. (Currently the test
fails, but it should be fixed when I merge in the latest changes from devel.)