Made ufuncs work with dask arrays
I used the simple fix of converting the dask arrays
underlying xarray Datasets/DataArrays to numpy arrays.

It might be more efficient to parallelize over chunks, but this
will require extra handling to make sure chunks align (e.g. in the
case of a basis function multiplying a footprint, since the basis
function isn't coming from a zarr object store, so it won't be chunked.)