Added log normal reparam and mu,sigma calc options
If `stdev` and optionally `mean` are passed as log normal parameters
then they will be converted to values of `mu` and `sigma` that give
that mean and stdev.

Additionally, the log normal can be reparameterised in terms of a normal
random variable if `reparameterise_log_normal = True` is passed in the
ini file.