Pass met_model as a list to get_footprint
`met_model` can be passed as list in an ini file, to
cover the case where two footprints are distiguished only
by met model. The list must have the same length as the number
of sites. Values of `None` will be ignored, so for sites where the
met model doesn't matter, `None` can be passed.