Added start/end to InvOut, fixes for time coords
Flux needs to use "flux_time" as a coordinate,
otherwise all the flux stats will be filled with
NaNs to match the times of the obs.

Added some post init steps in InversionOutput to do
this, as well as make sure that the basis in InversionOutput
has the same time coords as the flux.