Added pm.Data to track data used in model
Also a check to make sure `min_error` is the correct size.

Using `pm.Data` for external data used in the model will make sure
that data is added to the `InferenceData` returned from sampling, and
will apply dimensions and coordinates to that data.

This will make it easier to construct the outputs from the inversion.