Updated RHIME Likelihood function
The changes are:
1. instead of (Y - mu_bc)**2 in the uncertainty variable epsilon, we
have (Y - mu_bc)**1.5, which means that the distribution still has some
exponential decay (which in particular means the likelihood is "proper"
and has a finite integral)
2. the model-data mismatch uncertainty only has a factor proportional
to (Y - mu_bc) if Y > mu_bc (component-wise), since values of Y that are
below the baseline are not "pollution events"
3. instead of taking a maximum of min. model error and the combined obs
error + pollution event, the min model error is added directly. Since
the pollution event scaling only applies when Y > mu_bc, this term is
still cut off, but removing the maximum means that the likelihood has a
continuous derivative, since the derivative of the pollution event term
is 0 where at the point where it is cut off, and the min. model error
term has derivative 0.