Added guess at "time_period" attribute to flux
PARIS flux outputs need to know the flux frequency to make the correct
time offset.

The way this is done should probably be by passing the period
explicitly, but currently it is inferred, and for monthly inversions
using annual fluxes, the inferred time period is "1 year".

This PR fixes this by using the length of the inversion period to set
the "time period" attribute for flux data in some cases.