Make fp_sensitivity lazy
Removed `.as_numpy()`, which causes the H matrix to be computed.

This was done originally to avoid having a sparse arrays underlying the
dask array in the xr.DataArray.

Instead a new function was added that converts only the chunks from
sparse to dense.

This means that fp_sensitivity and bc_sensitivity are both lazy, so dask
can optimized their computation when fp_data is computed.